Add node objects to PowerElectronics module (#367)
* Add power electronics bus class

* also allows addings bus to system

* Automatically calculate system size

* Merge system allocate functions

* Update ScaleMicrogrid example

* Calculate internals/externals of the system

* System chooses internal variable mappings

* Inductor updated to take node input

* Update Resistor

* Update VoltageSource

* Update MicrogridBusDQ

* Update MicrogridLoad

* Update DistributedGenerator

* Update MicrogridLine

* Updated CHANGELOG.md

* Update DGTest example

* Update PowerElectronics enzyme example

* Rename DGSignal -> SignalNode

* Rename InfiniteBus -> GroundedBus

+++ GridKit/Model/PowerElectronics/CircuitComponent.hpp
@@ -72,7 +72,7 @@ namespace GridKit
     {
       if (connection_nodes_ == nullptr)
       {
-        connection_nodes_ = new IdxT[size_];
+        connection_nodes_ = new IdxT[static_cast<size_t>(size_)];
       }
 
       connection_nodes_[local_index] = global_index;
